rendering server errors
- changed Registration1 structure by removing RegistrationForm
- receiving error message from server and showing it on frontend if true
