+++ src/components/HeaderBurger.js
@@ -1,9 +1,10 @@
 import React from "react";
 import "../styles/HeaderBurger.css";
+import menu from "../img/menu-burger.svg";
 
 function HeaderBurger() {
     return (
-        <img className="headerBurger is-mobile" src="./src/img/menu-burger.svg" alt="Меню" />)
+        <img className="headerBurger is-mobile" src={menu} alt="Меню" />)
 };
 
 export default HeaderBurger;
