+++ src/components/Registration1/Registration1.css
@@ -16,3 +16,14 @@ hr {
         margin-top: 40px;
     }
 }
+
+.form-container {
+    width: 680px;
+    text-align: start;
+    margin: 0px auto;
+}
+@media screen and (max-width: 800px) {
+    .form-container {
+        width: 335px;
+    }
+}
