+++ src/components/RegistrationError/RegistrationError.css
@@ -1,7 +1,6 @@
 .reg_error {
     background-color: #EB5757;
     text-align: center;
-    display: none;
     justify-content: center;
     position: absolute;
     width: 100%;
