+++ src/components/RegistrationError/RegistrationError.js
@@ -1,13 +1,17 @@
-import React from "react";
+import React, {useState} from "react";
 import "./RegistrationError.css"
+import useServerError from "../../hooks/useServerError";
 
 function RegistrationError(props) {
 
-    return (
-        <div className={props.className} >
-            <p>{props.text}</p>
-        </div>
-    );
+   // const {serverError} = useServerError();
+
+        return (
+            <div className={props.className} style={props.style} >
+                <p>{props.text}</p>
+            </div>
+        );
+
 }
 
 export default RegistrationError;
