--- src/components/RegistrationForm/RegistrationForm.css
@@ -1,11 +1,0 @@
-.form-container {
-    width: 680px;
-    text-align: start;
-    margin: 0px auto;
-}
-@media screen and (max-width: 800px) {
-    .form-container {
-        width: 335px;
-    }
-}
-
