+++ src/index.js
@@ -3,6 +3,7 @@ import ReactDOM from "react-dom";
 import { BrowserRouter as Router } from "react-router-dom";
 import App from "./components/App.js";
 import Registration1 from "./components/Registration1/Registration1";
+import Test from "./components/Test";
